+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log (vs-script-commands)
 
+## 6.0.0 (November 19th, 2017; finished multi root support)
+
+* added `extension.scriptCommands.selectWorkspace` which can change between workspaces now
+
 ## 5.0.2 (October 14th, 2017; multi root support)
 
 * started to refactor to new, upcoming [Multi Root Workspace API](https://github.com/Microsoft/vscode/wiki/Extension-Authoring:-Adopting-Multi-Root-Workspace-APIs)